requirements.txt
africastalking==1.2.7
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contextlib2==21.6.0
coverage==6.5.0
coveralls==3.3.1
cryptography==42.0.3
Django==5.0.2
django-filter==23.5
django-nose==1.4.7
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
docopt==0.6.2
idna==3.6
jwcrypto==1.5.4
Markdown==3.5.2
nose==1.3.7
oauth2-provider==0.0
oauthlib==3.2.2
pycparser==2.21
python-coveralls==2.9.3
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
schema==0.7.5
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
